+++ packages/polaris-viz/CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Added
+
+- Added `hideLegendOverflow` prop to `<LineChart />` to all toggling of legend overflow behaviour
+
+### Changed
+
+- Changed `<DonutChart />` table legend width to only take as much space as the content requires.
 
 ## [14.1.2] - 2024-07-11